+++ webots_ros2_epuck/launch/robot_launch.py
@@ -27,6 +27,7 @@ from launch import LaunchDescription
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from ament_index_python.packages import get_package_share_directory
 from webots_ros2_driver.webots_launcher import WebotsLauncher
+from webots_ros2_driver.wait_for_controller_connection import WaitForControllerConnection
 from webots_ros2_driver.utils import controller_url_prefix
 
 
@@ -40,11 +41,11 @@ def get_ros2_nodes(*args):
     robot_description = pathlib.Path(os.path.join(package_dir, 'resource', 'epuck_webots.urdf')).read_text()
     ros2_control_params = os.path.join(package_dir, 'resource', 'ros2_control.yml')
     use_sim_time = LaunchConfiguration('use_sim_time', default=True)
+
+    # ROS control spawners
     controller_manager_timeout = ['--controller-manager-timeout', '50']
     controller_manager_prefix = 'python.exe' if os.name == 'nt' else ''
-
     use_deprecated_spawner_py = 'ROS_DISTRO' in os.environ and os.environ['ROS_DISTRO'] == 'foxy'
-
     diffdrive_controller_spawner = Node(
         package='controller_manager',
         executable='spawner' if not use_deprecated_spawner_py else 'spawner.py',
@@ -55,7 +56,6 @@ def get_ros2_nodes(*args):
             {'use_sim_time': use_sim_time},
         ],
     )
-
     joint_state_broadcaster_spawner = Node(
         package='controller_manager',
         executable='spawner' if not use_deprecated_spawner_py else 'spawner.py',
@@ -66,6 +66,7 @@ def get_ros2_nodes(*args):
             {'use_sim_time': use_sim_time},
         ],
     )
+    ros_control_spawners = [diffdrive_controller_spawner, joint_state_broadcaster_spawner]
 
     mappings = [('/diffdrive_controller/cmd_vel_unstamped', '/cmd_vel')]
     if 'ROS_DISTRO' in os.environ and os.environ['ROS_DISTRO'] in ['humble', 'rolling']:
@@ -125,22 +126,18 @@ def get_ros2_nodes(*args):
         }.items(),
     )
 
-    # Wait for the simulation to be ready to start the tools
-    tools_handler = launch.actions.RegisterEventHandler(
-        event_handler=launch.event_handlers.OnProcessExit(
-            target_action=diffdrive_controller_spawner,
-            on_exit=tool_nodes
-        )
+    # Wait for the simulation to be ready to start the tools and spawners
+    waiting_nodes = WaitForControllerConnection(
+        target_driver=epuck_driver,
+        nodes_to_start=[tool_nodes] + ros_control_spawners
     )
 
     return [
-        joint_state_broadcaster_spawner,
-        diffdrive_controller_spawner,
         robot_state_publisher,
         epuck_driver,
         footprint_publisher,
         epuck_process,
-        tools_handler,
+        waiting_nodes,
     ]