+++ webots_ros2_epuck/CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package webots_ros2_epuck
 2023.0.4 (2023-XX-XX)
 ------------------
 * Fixed ability to launch RViz without other tools.
+* Start ros control and navigation nodes when Webots is ready.
 
 2023.0.3 (2023-04-12)
 ------------------